bug fix to calstar.py
git-svn-id: svn+ssh://lyra.berkeley.edu/Bloom/amorgan/SVNREPO/q_soft@268 15f48517-4a6d-0410-94cc-8a7d56af1ecd

+++ trunk/Software/MiscBin/calstar.py
@@ -27,7 +27,8 @@ def magplot(reg, filelist):
     colornumber = len(callist)
     
     for index, star_reg in enumerate(callist):
-        os.remove(temppath)
+        if os.path.exists(temppath):
+            os.remove(temppath)
         datalist = []
         dataerrlist = []
         timelist = []
